requirements.txt
Flask
flask_sqlalchemy
pymysql
mysqldbmodel
pyservicebinding
python-dotenv
